+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Common impl for From<$w> for $t to automatically unwrap making API usage cleaner
 - `Account::deposit_of::<RESOURCE>` and `Account::withdraw_of::<RESOURCE>`
+- Comparison between ResourceDef and ResourceOf with == and !=
 ### Fixed
 - warnings on resource names (any case is now allowed)